Added all finished files to Github
All the code for the 2014-2015 year robot designed to stack and move
shipping crates around.  Code has large amounts of commented out
sections.  They were left in to explain how to use Pixy camera functions
and so the new programmer can see the debugging code that was used as
well as the integration for previous years and the usage of the command
group code.  Robot.java is the main file.
Lead Programmer: Garret Sampel